Change RunLoop to sleep after getting command
This should resolve some issue whereby the differing time taken to get a command between each time step will cause a lot of jitter in terms of the command send timing. The time taken to get a command is especially different when running the main controller because the main controller will skip some calculation sometimes.